Core and plugins separated, transloadit function exposed & an example app that uses to plugins
