--- scss/index.scss
@@ -1,6 +1,0 @@
-.transloadit-js-client {
-  font-family : "Comic Sans MS";
-  color       : purple;
-  border      : 1px dashed pink;
-  font-size   : 32px;
-}
