+++ src/js/lib/transloadit-dragndrop.js
@@ -1,0 +1,6 @@
+// The Dragndrop plugin
+
+export default function dragndrop(options) {
+  console.log(options.selector);
+  console.log('dragging and dropping here');
+}
