+++ src/js/lib/transloadit-instagram.js
@@ -1,0 +1,6 @@
+// The Dragndrop plugin
+
+export default function instagram(options) {
+  console.log(options);
+  console.log('Put a Filter On It (https://youtu.be/iHmLljk2t8M)!');
+}
